+++ pysat/instruments/methods/general.py
@@ -13,7 +13,6 @@
 import datetime as dt
 import numpy as np
 import pandas as pds
-import warnings
 
 import pysat