+++ pysat/tests/classes/cls_registration.py
@@ -16,7 +16,6 @@ Not directly called by pytest, but imported as part of test_registry.py.
 """
 
 import importlib
-import pytest
 import sys
 
 import pysat