+++ pysat/tests/test_instrument_index.py
@@ -9,10 +9,7 @@
 # ----------------------------------------------------------------------------
 """Unit tests for the `pysat.Instrument.index` attribute."""
 
-import datetime as dt
 from importlib import reload
-import numpy as np
-import warnings
 
 import pytest