+++ pysat/utils/__init__.py
@@ -6,20 +6,19 @@ This includes conversion of formats, loading of files, and user-supplied info
 for the pysat data directory structure.
 """
 
-from pysat.utils._core import available_instruments
-from pysat.utils._core import display_available_instruments
-from pysat.utils._core import display_instrument_stats
-from pysat.utils._core import generate_instrument_list
-from pysat.utils._core import get_mapped_value
-from pysat.utils._core import listify
-from pysat.utils._core import load_netcdf4
-from pysat.utils._core import NetworkLock
-from pysat.utils._core import scale_units
-from pysat.utils._core import stringify
-from pysat.utils._core import update_fill_values
-from pysat.utils import coords
-from pysat.utils import files
-from pysat.utils import io
-from pysat.utils import registry
-from pysat.utils import testing
-from pysat.utils import time
+from pysat.utils._core import available_instruments  # noqa: F401
+from pysat.utils._core import display_available_instruments  # noqa: F401
+from pysat.utils._core import display_instrument_stats  # noqa: F401
+from pysat.utils._core import generate_instrument_list  # noqa: F401
+from pysat.utils._core import get_mapped_value  # noqa: F401
+from pysat.utils._core import listify  # noqa: F401
+from pysat.utils._core import NetworkLock  # noqa: F401
+from pysat.utils._core import scale_units  # noqa: F401
+from pysat.utils._core import stringify  # noqa: F401
+from pysat.utils._core import update_fill_values  # noqa: F401
+from pysat.utils import coords  # noqa: F401
+from pysat.utils import files  # noqa: F401
+from pysat.utils import io  # noqa: F401
+from pysat.utils import registry  # noqa: F401
+from pysat.utils import testing  # noqa: F401
+from pysat.utils import time  # noqa: F401